requirements.txt
mcp
openai
pandas
matplotlib
seaborn
scikit-learn
streamlit
python-dotenv
pillow
torch
torchvision
